requirements.txt
APScheduler==3.6.3
attrs==21.4.0
autopep8==1.6.0
cachetools==4.2.2
certifi==2021.10.8
configparser==5.2.0
coverage==6.3.1
environs==9.5.0
greenlet==1.1.2
iniconfig==1.1.1
install==1.3.5
loguru==0.6.0
mariadb==1.0.11
marshmallow==3.14.1
packaging==21.3
pluggy==1.0.0
pockets==0.9.1
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.7
pytest==7.0.0
pytest-cov==3.0.0
pytest-timeout==2.1.0
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
tomli==2.0.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
